add machine clear.

+++ atomic.js
@@ -455,11 +455,11 @@ end
 
 					if ( initEngine = engctx.init )
 						ATOM.mixContext(sql, runctx.Entry, runctx, runctx => {  // mixin sql vars into engine query
-							//Log(">mix", engctx.thread, runctx);
+							Log(">mix", runctx);
 
 							if (runctx) 
 								initEngine(engctx.thread, engctx.code || "", runctx, err => {
-									//Log(">init", err);
+									Log(">init", err);
 									cb( err ? null : engctx );
 								});
 
@@ -478,11 +478,11 @@ end
 				Trace( `INIT ${engctx.thread} ON core${engctx.worker.id}`, req, Log );
 				
 				prime(req, engctx, engctx => {	// prime engine context
-					//Log(">prime", engctx);
+					Log(">prime", engctx);
 					
 					if (engctx) 
 						program(sql, engctx, engctx => {	// program/compile engine
-							// Log(">pgm", engctx);
+							Log(">pgm", engctx);
 							if (engctx)  // all went well so execute it
 								execute( engctx, cb );
 
@@ -592,9 +592,9 @@ end
 		 free/delete/DELETE.
 		*/
 			ATOM.run(req, (ctx,step) => {
-				Trace( `step ${ctx.thread}`, req, Log);
-				Log(">run", ctx);
+				Log(">init", ctx);
 				if ( ctx && step ) {
+					Trace( `step ${ctx.thread}`, req, Log);
 					for (var n=0, N=ctx.Runs||0; n<N; n++) step( ctx => {} );
 					res( ctx );
 				}
@@ -612,9 +612,14 @@ end
 		 free/delete/DELETE.
 		*/
 			ATOM.run(req, (ctx,step) => {
-				Trace( `kill ${ctx.thread}`, req, Log);
-				delete ATOM.context[ ctx.thread ];
-				res( ctx );
+				if (ctx) {
+					Trace( `kill ${ctx.thread}`, req, Log);
+					delete ATOM.context[ ctx.thread ];
+					res( ctx );
+				}
+				
+				else
+					res( null );
 			});
 		},
 
@@ -626,10 +631,11 @@ end
 		 free/delete/DELETE.
 		*/
 			ATOM.run( req, (ctx, step) => {  // get engine stepper and its context
-				Trace( `run ${ctx.thread}`, req, Log);
-				if ( ctx && step )
+				if ( ctx && step ) {
+					Trace( `run ${ctx.thread}`, req, Log);
 					step( ctx => res( ctx ) );
-				
+				}
+
 				else
 					res( null );
 			});
@@ -643,8 +649,13 @@ end
 		 free/delete/DELETE.
 		*/
 			ATOM.run( req, (ctx,step) => {
-				Trace( `init ${ctx.thread}`, req, Log);
-				res( ctx );
+				if ( ctx ) {
+					Trace( `init ${ctx.thread}`, req, Log);
+					res( ctx );
+				}
+				
+				else
+					res( null );
 			});
 		},
 
@@ -829,22 +840,22 @@ if 'PORT' in PORTS:
 		ERR = 103
 else:	# entry logic
 	if INIT:	#import global modules and connect to sqldb
-		try:
-			global IMP, JSON, SYS, FLOW, SQL0, SQL1, NP
-			import sys as SYS			#system info
-			import json as JSON			#json interface
-			from PIL import Image as IMP		#jpeg image interface
-			import mysql.connector as SQLC		#db connector interface
-			import numpy as NP
-			# import caffe as CAFFE		#caffe interface
-			# import flow as FLOW		# record buffering and loading logic
-			# setup sql connectors
-			SQL = SQLC.connect(user='${db.user}', password='${db.pass}', database='${db.name}')
-			# default exit codes and startup
-			ERR = 0
-			INIT = 0
-		except:
-			ERR = 107
+		global IMP, JSON, SYS, FLOW, SQL0, SQL1, NP
+		import sys as SYS			#system info
+		import json as JSON			#json interface
+		from PIL import Image as IMP		#jpeg image interface
+		import mysql.connector as SQLC		#db connector interface
+		import numpy as NP
+		# import caffe as CAFFE		#caffe interface
+		# import flow as FLOW		# record buffering and loading logic
+		# setup sql connectors
+		SQL = SQLC.connect(user='${db.user}', password='${db.pass}', database='${db.name}')
+		# default exit codes and startup
+		ERR = 0
+		INIT = 0
+		#try:
+		#except:
+		#	ERR = 107
 	else:
 		try:
 			# entry
